requirements.txt
python-dotenv==0.21.0
Unidecode==1.3.6
click~=8.1.6
paho-mqtt~=1.6.1
rich~=13.5.2
watchdog
